Move BaseQuadratic to base module
