+++ freeride/curves.py
@@ -1,155 +1,13 @@
 import numpy as np
 import matplotlib.pyplot as plt
-import numbers
 from freeride.plotting import textbook_axes, AREA_FILLS
-from freeride.formula import _formula, _quadratic_formula
-from freeride.base import PolyBase, QuadraticElement, AffineElement
+from freeride.formula import _formula
+from freeride.base import QuadraticElement, AffineElement, BaseQuadratic
 from IPython.display import Latex, display
 from bokeh.plotting import figure, show
 from bokeh.models import HoverTool, ColumnDataSource
 from freeride.exceptions import PPFError
 
-
-class BaseQuadratic:
-    '''
-    General piecewise quadratic objects.
-    '''
-
-    def __init__(self, intercept=None, linear_coef=None, quadratic_coef=None, elements=None):
-        """Create a piecewise quadratic curve.
-
-        Parameters
-        ----------
-        intercept : float or sequence of float, optional
-            Constant terms for each quadratic piece. If ``elements`` is not
-            provided these values are used to build :class:`QuadraticElement`
-            instances.
-        linear_coef : float or sequence of float, optional
-            Linear coefficients for each piece.
-        quadratic_coef : float or sequence of float, optional
-            Quadratic coefficients for each piece.
-        elements : list of :class:`QuadraticElement`, optional
-            Preconstructed elements defining the curve. When supplied the
-            coefficient arguments are ignored.
-
-        Notes
-        -----
-        Elements that do not define a domain are treated as active for all
-        quantities.
-        """
-        if elements is None:
-            if isinstance(linear_coef, (int, float)):
-                linear_coef = [linear_coef]
-            if isinstance(intercept, (int, float)):
-                intercept = [intercept]
-            if isinstance(quadratic_coef, (int, float)):
-                quadratic_coef = [quadratic_coef]
-            if (len(quadratic_coef) != len(intercept)) or (len(linear_coef) != len(intercept)):
-                raise ValueError("Coefficient lengths do not match.")
-
-            zipped = zip(intercept, linear_coef, quadratic_coef)
-            elements = [QuadraticElement(*coef) for coef in zipped]
-        self.elements = elements
-
-        if intercept is None:
-            intercept = [c.intercept for c in elements]
-        if linear_coef is None:
-            linear_coef = [c.linear_coef for c in elements]
-        if quadratic_coef is None:
-            quadratic_coef = [c.quadratic_coef for c in elements]
-
-        self.intercept = intercept
-        self.linear_coef = linear_coef
-        self.quadratic_coef = quadratic_coef
-
-    @classmethod
-    def from_points(cls, xy_points, fit=False):
-        """
-        Creates a Quadratic object from points.
-
-        Overdetermined systems are fit with least squares if `fit=True`. 
-        """
-        x = [i[0] for i in xy_points]
-        y = [i[1] for i in xy_points]
-        if fit:
-            coef = np.polyfit(x, y, 2)
-        else:
-            A_array = [[qp[0]**2, qp[0], 1] for qp in xy_points]
-            p_vals = [qp[1] for qp in xy_points]
-
-            A = np.array(A_array)
-            b = np.array(p_vals)
-            coef = np.linalg.solve(A, b)
-            
-        quad, lin, constant = tuple(coef)
-        return cls(constant, lin, quad)
-
-    @classmethod
-    def from_formula(cls, equation: str):
-        element = QuadraticElement.from_formula(equation)
-        return cls(elements = [element])
-
-    def horizontal_shift(self, delta, inplace=True):
-        new_elements = [e.horizontal_shift(delta, inplace=False) for e in self.elements]
-        if inplace:
-            self.__init__(elements=new_elements)
-        else:
-            return self.__class__(elements=new_elements)
-
-    def vertical_shift(self, delta, inplace=True):
-        new_elements = [e.vertical_shift(delta, inplace=False) for e in self.elements]
-        if inplace:
-            self.__init__(elements=new_elements)
-        else:
-            return self.__class__(elements=new_elements)
-
-    def plot(self, ax=None, set_lims=True, max_q=None, label=True, **kwargs):
-        """Plot the quadratic curve.
-
-        Parameters
-        ----------
-        ax : matplotlib.axes.Axes, optional
-            Axis on which to draw the plot. A new axis is created if omitted.
-        set_lims : bool, optional
-            When ``True`` the axes limits are adjusted based on the data.
-        max_q : float, optional
-            Maximum quantity value for the plot. Only used if the element has no
-            explicit domain.
-        label : bool, optional
-            If ``True`` annotate the axes with standard labels.
-        **kwargs : dict
-            Additional keyword arguments forwarded to ``matplotlib``.
-        """
-        if ax is None:
-            fig, ax = plt.subplots()
-
-        param_names = ['color', 'linewidth', 'linestyle', 'lw', 'ls']
-        plot_dict = {key: kwargs[key] for key in kwargs if key in param_names}
-        # Plot each element
-        for elmt in self.elements:
-            if elmt:
-                elmt.plot(ax=ax, label=label, max_q=max_q, **plot_dict)
-
-        if set_lims:
-            ax.relim()
-            ax.autoscale_view()
-
-        return ax
-
-    def active_element(self, q):
-        for piece in self.elements:
-            domain = piece._domain
-            if domain is None or (np.min(domain) <= q <= np.max(domain)):
-                return piece
-        return None
-
-    def __call__(self, q):
-        element = self.active_element(q)
-        if element:
-            return element(q)
-        else:
-            # q is out of range
-            return 0
 
 
 def intersection(element1, element2):
